Generate a custom Credo configuration
Fix a few style issues Credo was barking at to get the build passing.

At some point, we could move to `--strict` as well - the docs recommend that when using it as a linter, though I don’t know idiomatic Elixir style well enough to adjust the rules to match our current code.