Use custom analyzer for text_content
- The “english” analyzer removes stop words (“the”, “for”, “not”, etc.)
before indexing the content, which hurts us when searching for an exact
phrase (e.g. “Not for print”). ES now allows you to define an analyzer for
normal searching AND quote searching; so when doing a search, we are now
removing stop words for normal searches but leaving the stop words in place
when doing a quote search.