Handle query and root types more explicitly.
When preloading we don't really want the query or the top level type (eg
postStream and posts or editorialStream and editorial) to be in the
preload tree. However, if posts show up anywhere outside of the root we
do want to include them (eg posts in editorials or artist invites).